projects/discord-logger-bot/requirements.txt
discord.py
flask
python-dotenv
flask-cors
gunicorn
